% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSepalLength.r
\name{meanSepalLength}
\alias{meanSepalLength}
\title{Calculate mean sepal length for selectes species}
\usage{
meanSepalLength(iris, species = "setosa")
}
\arguments{
\item{iris}{iris default dataset}

\item{species}{select species of interest as character}
}
\value{
numeric value for mean of Sepal Length for selected species
}
\description{
Calculate mean sepal length for selectes species
}
