% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumThisColumn.r
\name{sumThisColumn}
\alias{sumThisColumn}
\title{Sum values in selected column}
\usage{
sumThisColumn(df, column)
}
\arguments{
\item{df}{df on interest}

\item{column}{single column number as numeric value}
}
\description{
Sum values in selected column
}
